sanphelp_microservice/private_message/requirements.txt
amqp==1.4.8
anyjson==0.3.3
Django==1.8.4
django-braces==1.8.1
django-cors-headers==1.1.0
django-extensions==1.5.7
django-filter==0.11.0
django-rest-swagger==0.3.5
django-storages==1.1.8
djangorestframework==3.2.4
enum34==1.0.4
drfdocs==0.0.6
Markdown==2.6.2
meld3==1.0.2
mysqlclient==1.3.10
MySQL-python==1.2.5
pdfkit==0.5.0
Pillow==2.9.0
pycrypto==2.6.1
PyJWT==1.4.0
pylint==1.5.6
pyparsing==1.5.7
pytz==2015.7
requests==2.7.0
supervisor==3.3.3
setproctitle==1.1.10	
six==1.9.0
xmltodict==0.11.0


